Add read handler to Postgres adapter; remove Controller
