+++ adapters/postgres/middleware/create.js
@@ -4,6 +4,7 @@ const _ = require('lodash');
 const handleQueryError = require('../util/handle-query-error');
 const formatTransaction = require('../util/format-transaction');
 const crud = require('../../../lib/sql/crud');
+const serverErrors = require('../../../lib/server-errors');
 
 module.exports = async function(req) {
   const pls = req.pls;
@@ -81,7 +82,7 @@ module.exports = async function(req) {
       Location: createdLink
     },
     body: {
-      data: formatTransaction(primaryTableResult, pls.definition, pls.version),
+      data: formatTransaction(primaryTableResult, pls.definition, pls.version, pls.adjustResourceQuantity),
       links: {
         self: createdLink
       }
