+++ server/resource/index.js
@@ -1,14 +1,10 @@
 'use strict';
 
-const Controller = require('./controller');
 const Routes = require('./routes');
 
-module.exports = function({definition, version, db, adapter}) {
-  var controller = new Controller({definition, version, db});
-
+module.exports = function({definition, version, adapter}) {
   var routes = new Routes({
     definition,
-    controller,
     version,
     adapter
   });
